GUI: Run application in the background
Moved most of the CLI handling code into NeovimQt::App and added
--nofork CLI option to nvim-qt.

In UNIX, if --nofork is missing, nvim-qt starts a new background
process as

    argv[0] --nofork [argv1 ...]

In Mac OSX and Windows the behaviour remains unchanged.
