+++ src/gui/app.h
@@ -5,15 +5,21 @@
 #include <QEvent>
 #include <QUrl>
 #include <QList>
+#include <QCommandLineParser>
 
 namespace NeovimQt {
 
+class NeovimConnector;
 class App: public QApplication
 {
 	Q_OBJECT
 public:
 	App(int &argc, char ** argv);
 	bool event(QEvent *event);
+	void showUi(NeovimConnector *c, const QCommandLineParser&);
+	static void processCliOptions(QCommandLineParser& p, const QStringList& arguments);
+	static NeovimConnector* createConnector(const QCommandLineParser& p);
+
 signals:
 	void openFilesTriggered(const QList<QUrl>);
 };
